split horizon select into component
